requirements.txt
absl-py==1.4.0
amqp==5.1.1
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
billiard==3.6.4.0
Brotli==1.0.9
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.32.0
cmake==3.26.3
colorlog==6.7.0
Cython==0.29.34
dnspython==2.3.0
email-validator==2.0.0.post2
facenet-pytorch==2.5.3
fastapi==0.95.1
filelock==3.12.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
Keras-Applications==1.0.8
keras-vggface @ git+https://github.com/nShravankg/Keras_Vggface.git@b6a1fcbc28e2ef92cf9b70db773460677f3c4cd8
kombu==5.2.4
libclang==16.0.0
lit==16.0.2
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
mtcnn==0.1.1
mutagen==1.46.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.11
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.17
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
vidgear==0.3.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.3.2
wrapt==1.14.1
yt-dlp==2023.3.4
